renamed & moved the base struct for active grotto class into its own file
